update balance, payment

+++ src/main/java/CustomerAccount.java
@@ -13,9 +13,8 @@ public class CustomerAccount extends Account{
     /** This field is the membership type that the customer is paying for */
     private Membership membership;
 
-    /** This field is the balance of the customer in their account */
-    //TODO make balance double
-    private int balance;
+    /** This field is the running total of money customer has spent paying for rides */
+    private double balance;
 
     // stack of ride ids user has
     private ArrayList<UUID> rideIdList;
@@ -83,7 +82,7 @@ public class CustomerAccount extends Account{
      * This is the setter method to update balance
      * @param balance is the new balance of the user
      */
-    void setBalance(int balance) {
+    void setBalance(double balance) {
         this.balance = balance;
     }
 
@@ -99,7 +98,7 @@ public class CustomerAccount extends Account{
      * This is the getter method to access the balance associated with the customer account
      * @return the balance associated with the customer account
      */
-    int getBalance() {
+    double getBalance() {
         return balance;
     }
